Fix 24 bugs: _os NameError, indicators cache, ADX NaN, circuit breaker reset, learning ratchet, blocking async calls, market regime, scoring clamp, and more

+++ budget.py
@@ -314,7 +314,7 @@ def get_budget_status() -> dict:
     }
 
 
-def calculate_position_size(price: float) -> int:
+def calculate_position_size(price: float) -> float:
     """Backward-compat alias."""
     qty, _ = compute_position_size(price)
     return qty
